pyproject.toml
[tool.poetry]
name = "pageloot-test"
version = "0.1.0"
description = "pageloot test task"
authors = ["Dmytro Novikov <novvikovdmitry@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
django-environ = "^0.11.2"
drf-yasg = "^1.21.8"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
mypy = "^1.13.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pre-commit = "^4.0.1"

[tool.black]
line-length = 88

[tool.flake8]
max-line-length = 88

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
